# dev/openapi.yaml
openapi: 3.0.3
info:
  title: Packing List API
  version: 0.1.0
  description: Backpacking gear management API
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/persons:
    get:
      tags:
        - persons
      summary: List all persons
      operationId: listPersons
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - persons
      summary: Create a new person
      operationId: createPerson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreate'
      responses:
        '201':
          description: Person created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/items:
    get:
      tags:
        - items
      summary: List all items
      operationId: listItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - items
      summary: Create a new item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/items/{itemId}:
    get:
      tags:
        - items
      summary: Get item by ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - items
      summary: Update item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - items
      summary: Delete item
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/sets:
    get:
      tags:
        - sets
      summary: List all item sets
      operationId: listSets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of item sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemSet'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - sets
      summary: Create a new item set
      operationId: createSet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemSetCreate'
      responses:
        '201':
          description: Item set created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemSet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/packs:
    get:
      tags:
        - packs
      summary: List all packs
      operationId: listPacks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pack'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - packs
      summary: Create a new pack
      operationId: createPack
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackCreate'
      responses:
        '201':
          description: Pack created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pack'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/trips:
    get:
      tags:
        - trips
      summary: List all trips
      operationId: listTrips
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - trips
      summary: Create a new trip
      operationId: createTrip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/trips/{tripId}:
    get:
      tags:
        - trips
      summary: Get trip by ID
      operationId: getTripById
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trip details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - trips
      summary: Update trip
      operationId: updateTrip
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdate'
      responses:
        '200':
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - trips
      summary: Delete trip
      operationId: deleteTrip
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Trip deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/trips/{tripId}/packs:
    get:
      tags:
        - trip-packs
      summary: List packs in a trip
      operationId: listTripPacks
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of packs in trip
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripPackWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - trip-packs
      summary: Add pack to trip
      operationId: addTripPack
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripPackCreate'
      responses:
        '201':
          description: Pack added to trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripPackWithDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/trips/{tripId}/packs/{packId}:
    delete:
      tags:
        - trip-packs
      summary: Remove pack from trip
      operationId: removeTripPack
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: packId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Pack removed from trip
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/trips/{tripId}/items:
    get:
      tags:
        - trip-items
      summary: List items in a trip
      operationId: listTripItems
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of items in trip
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripItemWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - trip-items
      summary: Add item to trip
      operationId: addTripItem
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripItemCreate'
      responses:
        '201':
          description: Item added to trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripItemWithDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/trips/{tripId}/sets/{setId}:
    delete:
      tags:
        - trip-sets
      summary: Remove set from trip
      operationId: removeTripSet
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: setId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Set removed from trip
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Person:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        birthdate:
          type: string
          format: date
          nullable: true
        body_weight_grams:
          type: number
          format: float
        conditioning_level:
          type: string
          enum: [sedentary, average, athletic, military]
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PersonCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        birthdate:
          type: string
          format: date
        body_weight_grams:
          type: number
          format: float
        conditioning_level:
          type: string
          enum: [sedentary, average, athletic, military]
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT